backend/app/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
firebase-admin==6.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
deepseek==1.0.0  # Replace with the actual DeepSeek package name
mysql-connector-python==8.1.0
requests==2.31.0
pytz==2023.3